refactor(linter): move shared function from utils to rule
